+++ FE/src/pages/Parents/History/HistoryDetailModal/HistoryCard/CardHeader/CardHeader.module.scss
@@ -1,0 +1,59 @@
+@import "@/styles/util.scss";
+@import "@/styles/color.scss";
+@import "@/styles/fontStyle.scss";
+
+.cardHeader {
+    width: 100%;
+    height: 66px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: 10px;
+    flex-shrink: 0;
+    padding-left: 16px;
+
+    .cardHeaderImgWrapper {
+        width: 42px;
+        height: 42px;
+        flex-shrink: 0;
+        border-radius: 50%;
+        border: 1px solid $primary;
+        background-color: $primary;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+
+        img {
+            width: 42px;
+            height: 42px;
+            border-radius: 50%;
+        }
+    }
+
+    .cardHeaderInfo {
+        width: max-content;
+        height: 100%;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: flex-start;
+        flex-grow: 1;
+
+        .cardHeaderInfoTitle {
+            @extend %SB_14;
+            color: $gray;
+        }
+
+        .cardHeaderInfoTime {
+            @extend %R_12;
+            color: $gray;
+        }
+    }
+
+    .report-btn {
+        @extend %M_14;
+        color: $gray;
+        height: 100%;
+        padding: 0 16px;
+    }
+}
