+++ FE/src/pages/Parents/History/HistoryDetailModal/HistoryCard/HistoryCard.module.scss
@@ -1,0 +1,29 @@
+@import "@/styles/color.scss";
+
+.card {
+    width: 90%;
+    height: max-content;
+    min-height: 465px;
+    flex-shrink: 0;
+    border-radius: 16px;
+    background-color: #fff;
+}
+
+.emptyHistory {
+    width: 100%;
+    height: 100%;
+    margin-right: 40px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 21px;
+
+    .emptyText {
+        font-size: 16px;
+        font-style: normal;
+        font-weight: 500;
+        line-height: 24px;
+        color: $gray;
+    }
+}
