+++ FE/src/pages/Parents/History/HistoryDetailModal/HistoryDetail.jsx
@@ -1,7 +1,8 @@
 import { Header } from "@/components/Header/Header";
 import styles from "./HistoryDetail.module.scss";
-import { useFetch } from "../../../../hooks/useFetch";
+import { useFetch } from "@/hooks/useFetch";
 import { useSearchParams } from "react-router-dom";
+import { EmptyHistoryCard, HistoryCard } from "./HistoryCard/HistoryCard";
 
 export function HistoryDetail({ driverProfile, closeModal }) {
     const [searchParams] = useSearchParams();
@@ -26,70 +27,6 @@ export function HistoryDetail({ driverProfile, closeModal }) {
             <main className={styles.cardList}>
                 {loading ? "loading..." : cardListElement}
             </main>
-        </div>
-    );
-}
-
-function EmptyHistoryCard() {
-    return (
-        <section className={styles.emptyHistory}>
-            <p className={styles.emptyText}>픽업 구독을 신청해보세요!</p>
-        </section>
-    );
-}
-
-function HistoryCard({ cardData, driverProfile: { driverImage, driverName } }) {
-    const { info } = cardData;
-    const isPickUpStart = true;
-    const getTime = (timeInfo) => {
-        const [date, time] = timeInfo.split("T");
-        return [date, time];
-    };
-    const [date, time] = isPickUpStart
-        ? getTime(info.startTime)
-        : getTime(info.endTime);
-
-    const kidImg = isPickUpStart ? info.startImage : info.endImage;
-
-    return (
-        <section className={styles.card}>
-            <CardHeader
-                date={date}
-                time={time}
-                driverImage={driverImage}
-                driverName={driverName}
-                status={info.status}
-            />
-            <CardMain img={kidImg} />
-        </section>
-    );
-}
-
-function CardHeader({ date, time, driverImage, driverName, status }) {
-    return (
-        <header className={styles.cardHeader}>
-            <div className={styles.cardHeaderImgWrapper}>
-                <img src={driverImage} alt="driver-profile" />
-            </div>
-            <div className={styles.cardHeaderInfo}>
-                <h4 className={styles.cardHeaderInfoTitle}>{driverName}</h4>
-                <span
-                    className={styles.cardHeaderInfoTime}
-                >{`${date} ${time} ${status}`}</span>
-            </div>
-        </header>
-    );
-}
-
-function CardMain({ img }) {
-    return (
-        <div className={styles.cardContent}>
-            <img
-                src={img}
-                alt="card-content"
-                className={styles.cardContentImg}
-            />
-            <p className={styles.cardContentText}>특이사항 없음</p>
         </div>
     );
 }
