+++ FE/src/pages/Parents/History/HistoryDetailModal/HistoryDetail.module.scss
@@ -21,116 +21,3 @@
     padding-top: 15px;
     gap: 19px;
 }
-
-.card {
-    width: 90%;
-    height: max-content;
-    min-height: 465px;
-    flex-shrink: 0;
-    border-radius: 16px;
-    background-color: #fff;
-}
-
-.cardHeader {
-    width: 100%;
-    height: 66px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    gap: 10px;
-    flex-shrink: 0;
-    padding-left: 16px;
-
-    .cardHeaderImgWrapper {
-        width: 42px;
-        height: 42px;
-        flex-shrink: 0;
-        border-radius: 50%;
-        border: 1px solid $primary;
-        background-color: $primary;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-
-        img {
-            width: 42px;
-            height: 42px;
-            border-radius: 50%;
-        }
-    }
-
-    .cardHeaderInfo {
-        width: max-content;
-        height: 100%;
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: flex-start;
-        flex-grow: 1;
-
-        .cardHeaderInfoTitle {
-            @extend %SB_14;
-            color: $gray;
-        }
-
-        .cardHeaderInfoTime {
-            @extend %R_12;
-            color: $gray;
-        }
-    }
-
-    .report-btn {
-        @extend %M_14;
-        color: $gray;
-        height: 100%;
-        padding: 0 16px;
-    }
-}
-
-.cardContent {
-    width: 100%;
-    height: max-content;
-}
-
-.cardContentImg {
-    width: 100%;
-    aspect-ratio: 1 / 1;
-}
-
-.cardContent {
-    width: 100%;
-    height: max-content;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-    gap: 16px;
-    img {
-        object-fit: cover;
-    }
-    .cardContentText {
-        @extend %R_14;
-        color: $gray;
-        padding: 0 12px;
-        width: 100%;
-        margin-bottom: 10px;
-    }
-}
-.emptyHistory {
-    width: 100%;
-    height: 100%;
-    margin-right: 40px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 21px;
-
-    .emptyText {
-        font-size: 16px;
-        font-style: normal;
-        font-weight: 500;
-        line-height: 24px;
-        color: $gray;
-    }
-}
